% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBDA_CleanUp.R
\name{CBDA_CleanUp}
\alias{CBDA_CleanUp}
\title{CBDA Clean up function for Compressive Big Data Analytics}
\usage{
CBDA_CleanUp(label = "CBDA_package_test", workspace_directory = tempdir())
}
\arguments{
\item{label}{This is the label appended to RData workspaces generated within the CBDA calls}

\item{workspace_directory}{Directory where the results and workspaces are saved}
}
\value{
value
}
\description{
This CBDA cleans the current directory where all the intermediate workspaces have been created.
}
