% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBDA_initialization.R
\name{CBDA_initialization}
\alias{CBDA_initialization}
\title{CBDA Initialization function for Compressive Big Data Analytics}
\usage{
CBDA_initialization(pkg = c("missForest", "stats", "utils", "prettydoc",
  "foreach", "SuperLearner", "knockoff", "caret", "smotefamily", "parallel",
  "doParallel", "glmnet"), install = FALSE)
}
\arguments{
\item{pkg}{List of packages to install and attach for running the CBDA algorithm.
A default list is already defined.}

\item{install}{Option to setup installation and attachment of the listed package.
Set to FALSE by default}
}
\value{
value
}
\description{
This CBDA function installs and attaches all the packages needed to run the CBDA.
 A user-defined list of packages can be passed as argument. Itis recommended to first execute
 the function without any arguments.
}
