% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBDA_spectrum_plots.R
\name{CBDA_spectrum_plots}
\alias{CBDA_spectrum_plots}
\title{CBDA Spectrum plot function for Compressive Big Data Analytics}
\usage{
CBDA_spectrum_plots(top)
}
\arguments{
\item{top}{Top ranked predictive models from the Learning/Training step}
}
\value{
value
}
\description{
This CBDA function generates histograms of the feature counts/densities as returned
 by the Accuracy and MSE metrics after the Learning/Training step.
}
